* lily/midi-item.cc (to_string): use to_int() iso. int cast.
* lily/staff-performer.cc (initialize): add tempo/name to
audio_staff_

+++ flower/rational.cc
@@ -286,7 +286,7 @@ Rational::to_string () const
 int
 Rational::to_int () const
 {
-  return num () / den ();
+  return (int) num () / den ();
 }
 
 int
